chore(frontend): 누적된 ESLint 에러 20건 정리 (#912)
저장소에 쌓여 있던 eslint 에러 20건을 정리해 `eslint .`를 0 errors로 되돌린다.

- 테스트 stub의 `as any`/`null as any` 제거 (정밀 타입), 미사용 변수·no-this-alias 정리
- react-refresh/only-export-components: WorkflowCanvas 상수/타입을 workflowCanvasData.ts로 분리
- 테스트가 없는 컴포넌트의 `no-explicit-any`(SlotEditForm·비밀번호 재설정 폼)와
  `set-state-in-effect`(WorkflowDetailPanel·EditWorkspaceDialog)는 동작·커버리지 보존을 위해
  실행 코드를 바꾸지 않고 사유를 명시한 eslint-disable로 처리

검증: eslint . 0 errors, tsc -b 0 errors, 영향 테스트 60건 통과
