+++ frontend/src/features/workflow-draft-read/ui/TransitionListPanel.test.tsx
@@ -18,16 +18,16 @@ const stubTransitions: WorkflowTransitionDetail[] = [
     to: "B",
     label: "조건",
     toPolicyRef: "POL_001",
-  } as any,
+  },
   {
     id: "edge-2",
     workflowDefinitionId: 10,
     domainPackVersionId: 3,
     from: "B",
     to: "C",
-    label: null as any,
-    toPolicyRef: null as any,
-  } as any,
+    label: undefined,
+    toPolicyRef: undefined,
+  },
 ];
 
 describe("TransitionListPanel", () => {
